Add support for the sim-folder.
The sim-folder contains files that are not tests themselves but rather
utilities that are simulation specific and might be publicly used by
other modules.

A use case is For example if you have a library of IP-cores that is
used by project you might want to test the project in isolation. In
this case you do not want to add the tests for the IP-cores when
testing the project but you still need their synth-files and
sim-files.