Filter get_cells commands in CDC modules to only return flop type cells.
This is done to reduce the risk of unintentionally ignoring paths to other
types of cells.

The PRIMITIVE_SUBGROUP property is now used to get 'flop' type cells.
For afifo.tcl, PRIMITIVE_GROUP==FLOP_LATCH was previously used as filter, which
failed for resync_counter during elaboration, because the primitive type is then
RTL_REGISTER. The 'flop' subgroup contains both FLOP_LATCH and RTL_REGISTER.
See UG912 for reference.