Do PYTHONPATH insert() instead of append() in examples
This will prefer local repo checkouts of the projects instead of packages installed with pip.
Previously when e.g. build.py was called in a local repo checkout, if tsfpga was also installed via pip the tsfpga package would be imported from the pip install rather than the local repo.