Remove undefined warning when converting std_logic
This printout appears a lot when simulating e.g. a FIFO, which it performs to_int(... and write_last).
Since write_last is not defined in many cases when write_valid is '0', this triggers the warning.
In one large testbench I inspected, the printouts resulted in a 35% execution time increase.
Hence it seems motivated to remove the printout.