Avoid using intersphinx in VHDL module documentation generator
Instead use hard coded links.

It should be possible to use the documentation builder in an offline environment.

Also, using intersphinx is not really reliable in the long term
Say that down the line, some label that we link to is removed from the website.
In that case it should still be possible to build documentation for old commits.
When using intersphinx this would not be possible, and documentation build would fail.